backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.24
pydantic>=2.4
motor>=3.3
python-dotenv>=1.0
orjson>=3.9
passlib[bcrypt]>=1.7
bcrypt>=4.0,<5
python-jose[cryptography]>=3.3
